drivers: gpu: Fix callback function type
Bug: 117299373

+++ drivers/gpu/msm/kgsl_pwrctrl.c
@@ -128,7 +128,7 @@ static void _record_pwrevent(struct kgsl_device *device,
 /**
  * kgsl_get_bw() - Return latest msm bus IB vote
  */
-static unsigned int kgsl_get_bw(void)
+static unsigned long kgsl_get_bw(void)
 {
 	return ib_votes[last_vote_buslevel];
 }
